+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 - Added [Watch-Process][] function
+- Refactor [gyo][] function
 
 ## [0.18.0] - 2026-03-08